server/requirements/prod.txt
alembic==1.14.1
fastapi==0.115.8
psycopg2==2.9.10
pydantic==2.10.6
python-dotenv==1.0.1
SQLAlchemy==2.0.38
uvicorn==0.34.0
